+++ lib/src/key/database/ti.rs
@@ -7,20 +7,20 @@ use serde::{Deserialize, Serialize};
 pub struct Ti {
 	__: u8,
 	_a: u8,
-	pub ns: u64,
+	pub ns: u32,
 	_b: u8,
-	pub db: u64,
+	pub db: u32,
 	_c: u8,
 	_d: u8,
 	_e: u8,
 }
 
-pub fn new(ns: u64, db: u64) -> Ti {
+pub fn new(ns: u32, db: u32) -> Ti {
 	Ti::new(ns, db)
 }
 
 impl Ti {
-	pub fn new(ns: u64, db: u64) -> Self {
+	pub fn new(ns: u32, db: u32) -> Self {
 		Ti {
 			__: b'/',
 			_a: b'+',
@@ -41,8 +41,8 @@ mod tests {
 		use super::*;
 		#[rustfmt::skip]
 		let val = Ti::new(
-			123u64,
-			234u64,
+			123u32,
+			234u32,
 		);
 		let enc = Ti::encode(&val).unwrap();
 		let dec = Ti::decode(&enc).unwrap();
